pyproject.toml
[project]
name = "clara"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.30.1",
    "joblib==1.5.2",
    "keras==3.10.0",
    "numpy==2.1",
    "pdf2image>=1.17.0",
    "pillow==11.3.0",
    "pymupdf>=1.26.4",
    "pytesseract==0.3.13",
    "scikit-learn==1.6.1",
    "sentence-transformers==5.1.0",
    "streamlit==1.49",
    "tensorflow==2.19.0",
    "tf-keras>=2.19.0",
]
